+++ example/models.py
@@ -185,3 +185,4 @@ class Company(models.Model):
 class Questionnaire(models.Model):
     name = models.CharField(max_length=100)
     questions = models.JSONField()
+    metadata = models.JSONField()