+++ rest_framework_json_api/schemas/openapi.py
@@ -681,9 +681,12 @@ class AutoSchema(drf_openapi.AutoSchema):
         and 'links'.
         """
         # TODO: remove attributes, etc. for relationshipView??
-        if isinstance(serializer.parent, serializers.ListField):
+        if isinstance(
+            serializer.parent, (serializers.ListField, serializers.BaseSerializer)
+        ):
             # Return plain non-JSON:API serializer schema for serializers nested inside
-            # a ListField, as those don't use the full JSON:API serializer schemas.
+            # a Serializer or a ListField, as those don't use the full JSON:API
+            # serializer schemas.
             return super().map_serializer(serializer)
 
         required = []